+++ backend/events/apps.py
@@ -5,7 +5,6 @@ import contextlib
 import time
 import ssl
 from django.apps import AppConfig
-
 from django.conf import settings
 
 import aio_pika
@@ -16,16 +15,7 @@ import orchestrator.common_pb2 as common_pb2
 import orchestrator.event_pb2 as event_pb2
 from substrapp.orchestrator import get_orchestrator_client
 
-from substrapp.tasks.tasks_prepare_task import prepare_task
 from substrapp.utils import get_owner
-
-from celery.result import AsyncResult
-
-from substrapp.tasks.tasks_compute_plan import delete_cp_pod_and_dirs_and_optionally_images
-from substrapp.tasks.tasks_remove_intermediary_models import (
-    remove_intermediary_models,
-    remove_intermediary_models_from_buffer
-)
 
 
 logger = structlog.get_logger('events')
@@ -42,12 +32,12 @@ def get_event_loop():
 
 
 def on_computetask_event(payload):
-    current_worker = get_owner()
-    worker_queue = f"{settings.ORG_NAME}.worker"
+    my_organisation = get_owner()
     asset_key = payload['asset_key']
     channel_name = payload['channel']
     event_kind = payload['event_kind']
     metadata = payload['metadata']
+    targeted_organisation = metadata['worker']
 
     log = logger.bind(
         asset_key=asset_key,
@@ -72,20 +62,21 @@ def on_computetask_event(payload):
                 models.extend(client.get_computetask_output_models(parent_key))
 
             model_keys = [model['key'] for model in models
-                          if model['owner'] == current_worker and client.can_disable_model(model['key'])]
+                          if model['owner'] == my_organisation and client.can_disable_model(model['key'])]
             if model_keys:
-                remove_intermediary_models.apply_async((channel_name, model_keys), queue=worker_queue)
+                from substrapp.tasks.tasks_remove_intermediary_models import queue_remove_intermediary_models_from_db
+                queue_remove_intermediary_models_from_db(channel_name, model_keys)
 
+            # Handle compute plan if necessary
             compute_plan = client.query_compute_plan(task['compute_plan_key'])
+
             if computeplan_pb2.ComputePlanStatus.Value(compute_plan['status']) in [
                     computeplan_pb2.PLAN_STATUS_DONE,
                     computeplan_pb2.PLAN_STATUS_CANCELED,
                     computeplan_pb2.PLAN_STATUS_FAILED]:
                 log.info('Compute plan finished', plan=compute_plan['key'], status=compute_plan['status'])
-                if not settings.DEBUG_KEEP_POD_AND_DIRS:
-                    delete_cp_pod_and_dirs_and_optionally_images.apply_async(
-                        (channel_name, compute_plan), queue=worker_queue
-                    )
+                from substrapp.tasks.tasks_compute_plan import queue_delete_cp_pod_and_dirs_and_optionally_images
+                queue_delete_cp_pod_and_dirs_and_optionally_images(channel_name, compute_plan=compute_plan)
 
     if event_task_status != computetask_pb2.STATUS_TODO:
         return
@@ -93,39 +84,32 @@ def on_computetask_event(payload):
     if event_pb2.EventKind.Value(event_kind) not in [event_pb2.EVENT_ASSET_CREATED, event_pb2.EVENT_ASSET_UPDATED]:
         return
 
-    if metadata['worker'] != current_worker:
-        log.info('Skipping task', current_worker=current_worker, task_worker=metadata["worker"])
+    if targeted_organisation != my_organisation:
+        log.info(
+            'Skipping task: this organisation is not the targeted organisation',
+            my_organisation=my_organisation,
+            targeted_organisation=targeted_organisation,
+            assert_key=asset_key,
+        )
         return
 
     with get_orchestrator_client(channel_name) as client:
         task = client.query_task(asset_key)
 
-    task_status = computetask_pb2.ComputeTaskStatus.Value(task["status"])
-
-    if task_status != event_task_status:
-        raise ValueError(f'Task {asset_key} status out of sync: task status={task["status"]} '
-                         f'!= event status={metadata["status"]}')
-
-    if AsyncResult(asset_key).state != 'PENDING':
-        log.info('Skipping task: already exists')
-        return
-
-    prepare_task.apply_async(
-        (channel_name, task),
-        task_id=asset_key,
-        queue=worker_queue
-    )
+    from substrapp.tasks.tasks_prepare_task import queue_prepare_task
+    queue_prepare_task(channel_name, task=task)
 
 
 def on_model_event(payload):
-    worker_queue = f"{settings.ORG_NAME}.worker"
     asset_key = payload["asset_key"]
     event_kind = payload['event_kind']
 
     logger.info('Processing model', asset_key=asset_key, kind=event_kind)
 
     if event_pb2.EventKind.Value(event_kind) == event_pb2.EVENT_ASSET_DISABLED:
-        remove_intermediary_models_from_buffer.apply_async([asset_key], queue=worker_queue)
+        from substrapp.tasks.tasks_remove_intermediary_models import remove_intermediary_models_from_buffer
+        # This task is broadcasted to all worker (see the broadcast defined in backend/celery.py)
+        remove_intermediary_models_from_buffer.apply_async([asset_key])
 
 
 async def on_message(message: aio_pika.IncomingMessage):
@@ -133,7 +117,6 @@ async def on_message(message: aio_pika.IncomingMessage):
         payload = json.loads(message.body)
         logger.debug("Received payload", payload=payload)
         asset_kind = common_pb2.AssetKind.Value(payload['asset_kind'])
-
         if asset_kind == common_pb2.ASSET_COMPUTE_TASK:
             on_computetask_event(payload)
         elif asset_kind == common_pb2.ASSET_MODEL: