+++ backend/substrapp/compute_tasks/volumes.py
@@ -28,7 +28,11 @@ def get_volumes(dirs: Directories, is_testtuple_eval: bool):
     volumes = [
         {
             "name": "subtuple",
-            "persistentVolumeClaim": {"claimName": settings.WORKER_PVC_SUBTUPLE},
+            "persistentVolumeClaim": {
+                "claimName": settings.WORKER_PVC_SUBTUPLE
+                if settings.WORKER_PVC_IS_HOSTPATH
+                else get_worker_subtuple_pvc_name()
+            },
         }
     ]
 
@@ -44,3 +48,11 @@ def _add(volume_mounts, task_dir: str, folder: str, read_only: bool = False):
             "readOnly": read_only,
         }
     )
+
+
+def get_worker_subtuple_pvc_name():
+    return f"{settings.WORKER_PVC_SUBTUPLE}-{os.getenv('HOSTNAME')}"
+
+
+def get_docker_cache_pvc_name():
+    return f"{settings.WORKER_PVC_DOCKER_CACHE}-{os.getenv('HOSTNAME')}"