+++ backend/substrapp/models/__init__.py
@@ -5,5 +5,7 @@ from .objective import Objective
 from .datamanager import DataManager
 from .algo import Algo
 from .model import Model
+from .computeplan_worker_mapping import ComputePlanWorkerMapping
 
-__all__ = ['DataSample', 'Objective', 'DataManager', 'Algo', 'Model']
+__all__ = ['DataSample', 'Objective', 'DataManager', 'Algo', 'Model',
+           'ComputePlanWorkerMapping']