+++ backend/substrapp/tests/tasks/tests_prepare_task.py
@@ -1,30 +1,65 @@
 import mock
-from unittest.mock import MagicMock
 from rest_framework.test import APITestCase
 from django.test import override_settings
-from django_celery_results.models import TaskResult
 from substrapp.tasks.tasks_prepare_task import prepare_task
 from orchestrator.client import OrchestratorClient
+from celery.exceptions import Ignore
 
 CHANNEL = "mychannel"
 
 
 @override_settings(LEDGER_CHANNELS={CHANNEL: {'chaincode': {'name': 'mycc'}, 'model_export_enabled': True}})
 class PrepareTaskTests(APITestCase):
+    def setUp(self):
+        self.task = {
+            "key": "subtuple_test",
+            "compute_plan_key": "flkey",
+            "status": "STATUS_TODO", "category":
+            "TASK_TRAIN"
+        }
+
+    def test_task_is_not_runnable(self):
+        doing_task = {
+            "key": "subtuple_test",
+            "compute_plan_key": "flkey",
+            "status": "STATUS_DOING",
+            "category": "TASK_TRAIN"
+        }
+        with mock.patch.object(OrchestratorClient, "query_task", return_value=doing_task), \
+                mock.patch(
+                    'celery.app.task.Task.request',
+                    new_callable=mock.PropertyMock) as mrequest, \
+                mock.patch(
+                    'celery.worker.request.Request.delivery_info',
+                    new_callable=mock.PropertyMock) as mdeliveryinfo:
+            mrequest.return_value = mdeliveryinfo
+            mdeliveryinfo.return_value = {"routing_key": "routing_key"}
+            # task is already being processed this should return without performing further operation
+            prepare_task(CHANNEL, self.task)
+
+    def test_prepare_tasks_todo_orchestrator_error(self):
+        with mock.patch.object(OrchestratorClient, "query_task", return_value=self.task), \
+                mock.patch('celery.app.task.Task.request', new_callable=mock.PropertyMock) as mrequest, \
+                mock.patch(
+                    'celery.worker.request.Request.delivery_info',
+                    new_callable=mock.PropertyMock) as mdeliveryinfo, \
+                mock.patch.object(OrchestratorClient, "update_task_status") as mtask_status:
+            mrequest.return_value = mdeliveryinfo
+            mdeliveryinfo.return_value = {"routing_key": "routing_key"}
+            mtask_status.side_effect = Exception("orchestrator error")
+            with self.assertRaises(Ignore):
+                prepare_task(CHANNEL, self.task)
+
     def test_prepare_tasks(self):
-
-        task = {"key": "subtuple_test", "compute_plan_key": "flkey", "status": "STATUS_TODO"}
-
-        with mock.patch.object(TaskResult.objects, "filter") as mtaskresult, mock.patch(
-            "substrapp.tasks.tasks_prepare_task.json.loads"
-        ):
-
-            mock_filter = MagicMock()
-            mock_filter.count.return_value = 1
-            mtaskresult.return_value = mock_filter
-
-            with mock.patch.object(OrchestratorClient, "query_task", return_value=task), \
-                 mock.patch("substrapp.tasks.tasks_prepare_task.compute_task.apply_async",
-                            return_value="compute_task"), \
-                 mock.patch.object(OrchestratorClient, "update_task_status"):
-                prepare_task(CHANNEL, task)
+        with mock.patch.object(OrchestratorClient, "query_task", return_value=self.task), \
+                mock.patch("substrapp.tasks.tasks_prepare_task.compute_task.apply_async",
+                           return_value="compute_task") as mapply_async, \
+                mock.patch('celery.app.task.Task.request', new_callable=mock.PropertyMock) as mrequest, \
+                mock.patch(
+                    'celery.worker.request.Request.delivery_info',
+                    new_callable=mock.PropertyMock) as mdeliveryinfo, \
+                mock.patch.object(OrchestratorClient, "update_task_status"):
+            mrequest.return_value = mdeliveryinfo
+            mdeliveryinfo.return_value = {"routing_key": "routing_key"}
+            mapply_async.return_value = "compute_task"
+            prepare_task(CHANNEL, self.task)